# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REST API Pedro",
        "description": "REST API -> TCC of Pedro Rodrigues Pereira",
        "contact": {
            "email": "jierlausson@hotmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/units": {
            "get": {
                "tags": [
                    "Units"
                ],
                "summary": "List all units",
                "operationId": "a57391dfe211a9ccc8c604ddacf0bdf3",
                "responses": {
                    "200": {
                        "description": "List of units",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Unit"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Units"
                ],
                "summary": "Create a new unit",
                "operationId": "5a978a16bf0d0521e351d3b5f3a39b5e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Unit created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unit"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/units/{unit}": {
            "get": {
                "tags": [
                    "Units"
                ],
                "summary": "Retrieve a specific unit",
                "operationId": "5cc6969e4da70828e852596deee69619",
                "parameters": [
                    {
                        "name": "unit",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unit details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unit"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Units"
                ],
                "summary": "Update a unit",
                "operationId": "e324211c54695b6cc6b8d0447a2568a1",
                "parameters": [
                    {
                        "name": "unit",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Unit updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unit"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Units"
                ],
                "summary": "Delete a unit",
                "operationId": "2bde479ef6d3c6c15d8d6eec9edff638",
                "parameters": [
                    {
                        "name": "unit",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unit deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Unit": {
                "title": "Unit",
                "description": "Modelo da unidade",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID da unidade",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nome da unidade",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descrição da unidade",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}